requirements.txt
# Flask for web server
flask>=2.3.3
python-dotenv>=1.0.0

# Audio capture
sounddevice>=0.4.6

# Speech-to-Text
openai-whisper>=20230918
webrtcvad>=2.0.10

# Neural TTS (Coqui)
TTS>=0.17.6

# Core dependencies
numpy>=1.24.3
requests>=2.31.0
openai>=1.9.0

# Async and streaming support
anyio>=3.7.1

# MCP SDK
mcp[cli]>=1.7.0
